+++ HW5_Coroutines_Compose/app/src/main/java/com/example/hw5_coroutines_compose/Constants.kt
@@ -1,0 +1,5 @@
+package com.example.hw5_coroutines_compose
+
+object Constants {
+    const val PACKAGE = "package"
+}
