+++ HW5_Coroutines_Compose/app/src/main/res/values/strings.xml
@@ -3,7 +3,7 @@
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="enter_coroutines_number">Введите число корутин...</string>
-    <string name="choose_thread_pool_type">Выберите тип пула потоков</string>
+    <string name="choose_thread_pool_type">Выберите тип диспетчера</string>
     <string name="start_coroutines">Запуск корутин</string>
     <string name="stop_coroutines">Остановка корутин</string>
     <string name="notifications_allowed">Уведомления разрешены</string>
@@ -20,6 +20,10 @@
     <string name="coroutines_cancelled">Отменено %d корутин</string>
     <string name="coroutine_started">Корутина %1$d запущена</string>
     <string name="coroutine_finished">Корутина %1$d завершена</string>
+    <string name="notification_permission_title">Разрешение на уведомления</string>
+    <string name="notification_permission_message">Для отправки уведомлений необходимо предоставить разрешение.</string>
+    <string name="open_settings_button">Открыть настройки</string>
+    <string name="package_name">package</string>
 
 
     <string-array name="coroutines_thread_types">
